backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0
pydantic-settings==2.7.0
google-cloud-firestore==2.20.0
google-generativeai==0.8.0
google-api-python-client==2.160.0
google-auth==2.38.0
httpx==0.28.0
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
